seeding and factories for Bank Transactions and Entities

+++ billie/database/migrations/2018_12_06_151933_create_bank_transactions_table.php
@@ -16,7 +16,7 @@ class CreateBankTransactionsTable extends Migration
         Schema::create('bank_transactions', function (Blueprint $table) {
             $table->increments('id');
             $table->uuid('uuid')->unique();
-            $table->float('amount')->default(0);
+            $table->double('amount')->default(0);
             $table->dateTime('booking_date');
 
             // probably not needed but I leaving them here.
